feat(bot): send message about chat sale
